+++ lib/lines.h
@@ -1,5 +1,5 @@
 #include "pixel_array.h"
 
-void draw_horizontal_line(PixelArray* pixel_array,size_t x1,size_t x2,size_t y,uint32_t color);
-void draw_vertical_line(PixelArray* pixel_array,size_t x,size_t y1,size_t y2,uint32_t color);
-void draw_line(PixelArray* pixel_array,size_t x1,size_t y1,size_t x2,size_t y2,uint32_t color);
+void draw_horizontal_line(PixelArray* pixel_array,uint32_t x1,uint32_t x2,uint32_t y,uint32_t color);
+void draw_vertical_line(PixelArray* pixel_array,uint32_t x,uint32_t y1,uint32_t y2,uint32_t color);
+void draw_line(PixelArray* pixel_array,uint32_t x1,uint32_t y1,uint32_t x2,uint32_t y2,uint32_t color);
